requirements.txt
# Flask and web framework
Flask==3.1.0
cloudinary==1.44.0
requests==2.32.3
python-dotenv==1.1.0

# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # For YOLOv9

# Experiment tracking and pipeline management
clearml>=1.12.0

# Hyperparameter optimization
optuna>=3.3.0

# Data processing and visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
tqdm>=4.65.0
PyYAML>=6.0.0
pathlib>=1.0.1

# Optional: GPU support
# torch-cuda  # Uncomment if using CUDA 